requirements.txt
bottle==0.12.19
bottle-websocket==0.2.9
Eel==0.14.0
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
pyparsing==2.4.7
speedtest-cli==2.1.3
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
